refactor main dart and extract authWraper
